refactor: extract safe load with base resolver to a util
