+++ sdk/ml/azure-ai-ml/azure/ai/ml/_internal/__init__.py
@@ -1,8 +1,8 @@
 # ---------------------------------------------------------
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # ---------------------------------------------------------
-# pylint: disable=unused-import
-from ._util import enable_internal_components_in_pipeline
+
+from ._setup import enable_internal_components_in_pipeline
 from .entities import (
     Ae365exepool,
     AISuperComputerConfiguration,
