+++ sdk/ml/azure-ai-ml/azure/ai/ml/_internal/_utils/__init__.py
@@ -1,0 +1,9 @@
+# ---------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# ---------------------------------------------------------
+
+from ._yaml_utils import yaml_safe_load_with_base_resolver
+
+__all__ = [
+    "yaml_safe_load_with_base_resolver"
+]
